Fix Vault Forbidden bug
